requirements.txt
aggdraw==1.3.19
beautifulsoup4==4.14.3
google_api_python_client==2.192.0
google_auth_oauthlib==1.3.0
google-cloud-secret-manager==2.18.0
groq==1.0.0
moviepy==2.2.1
numpy==2.2.6
openai==2.14.0
Pillow==11.3.0
protobuf==4.25.3
pydub==0.25.1
pysrt==1.1.2
python-dotenv==1.2.2
pytube==15.0.0
Requests==2.32.5
google-search-results==2.4.2
torch==2.9.1
transformers==4.57.3
ultralytics==8.4.21
faster-whisper==1.1.1
yt_dlp==2026.3.3
opencv-python==4.12.0.88
fal-client